+++ crates/jcode-desktop2/src/editor.rs
@@ -239,10 +239,9 @@ impl Editor {
 
     /// Ctrl+W / Alt+Backspace: delete the word before the cursor.
     pub fn delete_word_back(&mut self) {
-        // An active selection is replaced, matching delete_back/delete_forward
-        // and normal text-field behavior. Without this the drain below shortens
-        // the buffer while `anchor` keeps pointing past the new end, and the
-        // next Backspace slices out of bounds (#728).
+        // Replace an active selection, like delete_back does. Otherwise the
+        // drain below leaves `anchor` past the new end and the next Backspace
+        // slices out of bounds (#728).
         if self.delete_selection().is_some() {
             return;
         }
@@ -764,50 +763,6 @@ mod tests {
         editor.set_cursor(0);
         editor.delete_word_forward();
         assert_eq!((editor.text(), editor.cursor()), ("beta", 0));
-    }
-
-    /// Regression for #728: a word-delete with an active selection used to
-    /// shorten the buffer while leaving `anchor` pointing past the new end.
-    /// The next ordinary Backspace then sliced out of bounds and panicked
-    /// across winit's Objective-C `key_down` boundary, aborting the process.
-    #[test]
-    fn word_delete_with_a_selection_does_not_leave_a_stale_anchor() {
-        // The reporter's exact sequence: type "hello world", Shift+Left,
-        // Option+Backspace, Backspace.
-        let mut editor = Editor::with_text("hello world");
-        editor.extend_left();
-        assert!(editor.selection().is_some(), "expected an active selection");
-
-        editor.delete_word_back();
-        assert!(
-            editor.selection().is_none(),
-            "word-delete must consume the selection, not leave a stale anchor"
-        );
-        assert!(
-            editor.cursor() <= editor.text().len(),
-            "cursor {} past buffer {:?}",
-            editor.cursor(),
-            editor.text()
-        );
-
-        // Must not panic.
-        editor.delete_back();
-        assert!(editor.cursor() <= editor.text().len());
-    }
-
-    #[test]
-    fn word_delete_forward_with_a_selection_replaces_the_selection() {
-        let mut editor = Editor::with_text("hello world");
-        editor.set_cursor(0);
-        editor.extend_right();
-        assert!(editor.selection().is_some());
-
-        editor.delete_word_forward();
-        assert_eq!(editor.text(), "ello world");
-        assert!(editor.selection().is_none());
-
-        editor.delete_back();
-        assert!(editor.cursor() <= editor.text().len());
     }
 
     #[test]
