+++ scripts/code_size_budget.json
@@ -7,7 +7,7 @@
     "crates/jcode-app-core/src/server/client_lifecycle.rs": 3180,
     "crates/jcode-app-core/src/server/client_session.rs": 1710,
     "crates/jcode-app-core/src/server/comm_control.rs": 2625,
-    "crates/jcode-app-core/src/server/comm_session.rs": 1433,
+    "crates/jcode-app-core/src/server/comm_session.rs": 1435,
     "crates/jcode-app-core/src/server/debug_server_state.rs": 1257,
     "crates/jcode-app-core/src/server/jade_relay.rs": 1429,
     "crates/jcode-app-core/src/server/provider_control.rs": 1600,
@@ -34,7 +34,7 @@
     "crates/jcode-base/src/skill.rs": 1426,
     "crates/jcode-base/src/todo.rs": 1277,
     "crates/jcode-config-types/src/lib.rs": 1496,
-    "crates/jcode-desktop2/src/editor.rs": 1364,
+    "crates/jcode-desktop2/src/editor.rs": 1373,
     "crates/jcode-desktop2/src/main.rs": 1667,
     "crates/jcode-desktop2/src/scene.rs": 1256,
     "crates/jcode-desktop2/src/states.rs": 1243,
